Resolve target helper merge conflicts
