+++ app/core/session_state.py
@@ -5,7 +5,8 @@ from __future__ import annotations
 import pandas as pd
 import streamlit as st
 
-from app.config import ALL_COLUMNS, DEFAULT_TARGETS, get_target_weights, normalise_target_weights
+from app.config import ALL_COLUMNS
+from app.core.targets import DEFAULT_TARGETS, get_target_weights, normalise_target_weights
 
 DEFAULT_WEIGHT_COLUMNS = ["Date", "Poids (Kgs)"]
