+++ app/pages/Settings.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 import streamlit as st
 
-from app.config import AppDefaults, DUPLICATE_STRATEGIES, get_target_weights, normalise_target_weights
+from app.config import AppDefaults, DUPLICATE_STRATEGIES
+from app.core.targets import get_target_weights, normalise_target_weights
 
 
 def main() -> None:
